update:增加lyrical编译

+++ script/linux_jazzy/build_docker_env.sh
@@ -39,6 +39,26 @@ if [[ -z ${ros2_version} ]]; then
   echo "设置默认的ros2版本: ${ros2_version}"
 fi
 
+# 先删除目录
+rm -rf ${ros2_dir}
+mkdir -p ${ros2_dir}
+
+# 阻止交互式命令 使用默认的
+export DEBIAN_FRONTEND=noninteractive
+
+apt update -y
+apt dist-upgrade -y
+
+# 时区
+apt -y install tzdata
+ln -fs /usr/share/zoneinfo/Asia/Shanghai /etc/localtime
+dpkg-reconfigure -f noninteractive tzdata
+
+# https://docs.ros.org/en/jazzy/Installation/Alternatives/Ubuntu-Development-Setup.html
+
+apt install curl wget xz-utils lsb-release locales -y
+
+echo "$(lsb_release -si)"
 if lsb_release -si 2>/dev/null | grep -qi ubuntu; then
     OS="Ubuntu"
 else
@@ -55,23 +75,8 @@ else
 fi
 echo "当前ros2版本:${ros2_version}"
 
-apt update -y
-apt dist-upgrade -y
-
-# 阻止交互式命令 使用默认的
-export DEBIAN_FRONTEND=noninteractive
-
-# 时区
-apt -y install tzdata
-ln -fs /usr/share/zoneinfo/Asia/Shanghai /etc/localtime
-dpkg-reconfigure -f noninteractive tzdata
-
-# https://docs.ros.org/en/jazzy/Installation/Alternatives/Ubuntu-Development-Setup.html
 
 echo "设定locale"
-
-apt install locales -y
-
 echo "=== 检查当前locale设置 ==="
 locale
 
@@ -93,10 +98,10 @@ export LANG=en_US.UTF-8
 export LC_ALL=en_US.UTF-8
 export LANGUAGE=en_US.UTF-8
 
-# 持久化
-echo "LANG=en_US.UTF-8" > /etc/default/locale
-echo "LC_ALL=en_US.UTF-8" >> /etc/default/locale
-echo "LANGUAGE=en_US.UTF-8" >> /etc/default/locale
+# 持久化 docker中没用
+# echo "LANG=en_US.UTF-8" > /etc/default/locale
+# echo "LC_ALL=en_US.UTF-8" >> /etc/default/locale
+# echo "LANGUAGE=en_US.UTF-8" >> /etc/default/locale
 
 echo "=== 当前locale设置 ==="
 locale
@@ -108,18 +113,21 @@ if [ "$OS" = "Ubuntu" ]; then
   add-apt-repository universe -y
 fi
 
-# 重来一次
-apt update -y
-
-apt install curl -y
-
 export ROS_APT_SOURCE_VERSION=$(curl -s https://api.github.com/repos/ros-infrastructure/ros-apt-source/releases/latest | grep -F "tag_name" | awk -F'"' '{print $4}')
-curl -L -o ${work_dir}/ros2-apt-source.deb "https://github.com/ros-infrastructure/ros-apt-source/releases/download/${ROS_APT_SOURCE_VERSION}/ros2-apt-source_${ROS_APT_SOURCE_VERSION}.$(. /etc/os-release && echo ${UBUNTU_CODENAME:-${VERSION_CODENAME}})_all.deb"
+apt_source_url="https://github.com/ros-infrastructure/ros-apt-source/releases/download/${ROS_APT_SOURCE_VERSION}/ros2-apt-source_${ROS_APT_SOURCE_VERSION}.$(. /etc/os-release && echo ${UBUNTU_CODENAME:-${VERSION_CODENAME}})_all.deb"
+echo "ros2-apt-source.deb下载地址:${apt_source_url}"
+curl -L -o ${work_dir}/ros2-apt-source.deb "${apt_source_url}"
+ls -alh
+if [ ! -f "${work_dir}/ros2-apt-source.deb" ]; then
+  echo "错误：软件包下载失败，请检查网络或系统兼容性。"
+  exit 1
+fi
 dpkg -i ${work_dir}/ros2-apt-source.deb
 
-apt update -y
 
 echo "安装开发工具: $(lsb_release -r)"
+# 这里必须更新 ros2-apt-source.deb 这个更新了源
+apt update -y
 apt install -y \
   python3-flake8-blind-except \
   python3-flake8-class-newline \
@@ -160,7 +168,6 @@ rm -rf ${llvm_path}/*.tar.xz
 
 
 echo "下载源码"
-rm -rf ${ros2_dir}
 mkdir -p ${ros2_dir}/src
 cd ${ros2_dir}
 
@@ -168,11 +175,11 @@ src_addr="https://raw.githubusercontent.com/ros2/ros2/${ros2_version}-release/ro
 echo "源码地址:${src_addr}"
 vcs import --input ${src_addr} src
 
-# Install colcon mixins
+echo "安装colcon mixins"
 colcon mixin add default https://github.com/colcon/colcon-mixin-repository/raw/master/index.yaml
 colcon mixin update default
 
-# 使用 rosdep 安装依赖
+echo "使用rosdep安装依赖"
 rosdep init
 rosdep update
 
@@ -205,15 +212,6 @@ fi
 
 echo "安装依赖"
 rosdep install --from-paths src --ignore-src -y --skip-keys "fastcdr rti-connext-dds-6.0.1 urdfdom_headers"
-
-
-# dest_arch="amd64"
-# if [[ -n ${cpuinfo} ]]; then
-#   dest_arch="arm64"
-# fi
-# echo "编译平台:${dest_arch}"
-# # 替换平台
-# sed -i "s/^.*CMAKE_SYSTEM_PROCESSOR.*$/set(CMAKE_SYSTEM_PROCESSOR ${dest_arch})/" ${work_dir}/toolchain.cmake
 
 
 apt-get clean
